mem0Client/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
Pillow>=10.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
bcrypt>=4.0.0
cryptography>=40.0.0
mem0ai
openai>=1.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
click>=8.0.0
rich>=13.0.0
python-dateutil>=2.8.0
tenacity>=8.0.0
PyJWT>=2.8.0
psycopg2-binary>=2.9.0